feat: badge 기능 구현
